requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0

# LLM integrations (optional - install based on provider choice)
openai>=1.0.0
anthropic>=0.7.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: For better HTML parsing if needed
lxml>=4.9.0